Add apis UserListAPI, UserRetrieveAPI, UserChatListAPI, UserChatCreateAPI
