+++ Pipfile
@@ -18,3 +18,5 @@ Markdown = "==3.2.1"
 django-crispy-forms = "*"
 isort = "*"
 django-cors-headers = "*"
+pylint = "*"
+flake8 = "*"
