+++ core/serializers/user.py
@@ -1,0 +1,16 @@
+# core/serializers/user.py
+
+from django.contrib.auth.models import User
+from rest_framework import serializers
+
+
+class UserListSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = ('id', 'username')
+
+
+class UserRetrieveSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = ('id', 'username')
