+++ core/views/group.py
