+++ templates/header.html
@@ -12,9 +12,6 @@
             <a class="navbar-brand" href="{% url 'home' %}">
                 <i class="fa fa-comments"></i>Chat
             </a>
-            <a class="navbar-brand" href="{% url 'groups' %}">
-                <i class="fa fa-comments"></i>Groups
-            </a>
         </div>
 
         <div id="navbar" class="navbar-collapse collapse">
